Handle rare error case where "u" message arrives before "b"
The original ch.py also does not handle this error case. Every
time the server reports a new message it sends a raw message
starting with "b" followed by a confirmation message starting
with "u".  In rare cases, the server sends the "u" message first
which previously caused the message to be missed by the bot.